Feature/content search indexer migration segregation (#2577)
* Split up migration script

* Fixed format

* Fixed README file
