+++ apps/services/search-indexer/src/migrate/migrateElastic.ts
@@ -1,59 +1,16 @@
-import yargs from 'yargs'
 import { logger } from '@island.is/logging'
 import { environment } from '../environments/environment'
-import * as aws from './aws'
-import * as dictionary from './dictionary'
-import * as elastic from './elastic'
-import * as kibana from './kibana'
+import * as aws from './lib/aws'
+import * as dictionary from './lib/dictionary'
+import * as elastic from './lib/elastic'
 import * as indexManager from '@island.is/content-search-index-manager'
 
 const { locales } = environment
 
 class App {
   async run() {
-    logger.info('Starting migration of dictionaries and ES config', environment)
-
-    const hasAwsAccess = await aws.checkAWSAccess()
-
-    if (hasAwsAccess) {
-      await this.migrateAws()
-    }
-
-    await this.migrateES(hasAwsAccess)
-
-    try {
-      await this.migrateKibana()
-    } catch (e) {
-      logger.error('Failed migrating kibana', e)
-    }
-
-    logger.info('Done!')
-  }
-
-  private async migrateAws(): Promise<boolean> {
-    logger.info('Starting aws migration')
-
-    const dictionaryVersion = indexManager.getDictionaryVersion() // returns versions of the dictionary in order with the newest version first
-
-    logger.info('Requesting files for dictionary version', {
-      version: dictionaryVersion,
-    })
-
-    const versionsDictionaryFiles = await dictionary.getDictionaryFilesForVersion(
-      dictionaryVersion,
-    )
-
-    // we will always validate packages against (s3 -> AWS ES -> AWS ES search domain) to ensure we don't have partial updates
-    logger.info('Starting validation of S3 and AWS packages')
-    const s3Files = await aws.uploadS3DictionaryFiles(versionsDictionaryFiles) // upload repo files to s3
-    const newEsPackages = await aws.createAwsEsPackages(s3Files) // create the dictionary packages files in AWS ES
-    await aws.associatePackagesWithAwsEsSearchDomain(newEsPackages) // attach the new packages to our AWS ES search domain
-    logger.info('All S3 and AWS packages up to date, AWS migration complete')
-    return true
-  }
-
-  private async migrateES(hasAwsAccess: boolean): Promise<boolean> {
     logger.info('Starting elasticsearch migration')
+    const hasAwsAccess = await aws.checkAWSAccess()
 
     // get dictionary AWS ES packages
     let esPackages: aws.AwsEsPackage[]
@@ -149,29 +106,11 @@ class App {
     logger.info('Elasticsearch migration completed')
     return true
   }
-
-  private async migrateKibana() {
-    logger.info('Starting kibana migration')
-    const version = indexManager.getElasticVersion()
-    await kibana.importObjects(version)
-    logger.info('Done')
-  }
-
-  async syncKibana() {
-    logger.info('Starting kibana syncing')
-    await kibana.syncObjects()
-    logger.info('Done')
-  }
 }
 
 async function migrateBootstrap() {
-  const argv = yargs(process.argv).argv
   const app = new App()
-  if (argv.syncKibana) {
-    await app.syncKibana()
-  } else {
-    await app.run()
-  }
+  await app.run()
 }
 
 migrateBootstrap().catch((error) => {
