+++ apps/services/search-indexer/webpack.config.js
@@ -1,5 +1,12 @@
+const path = require('path')
 module.exports = (config) => {
-  config.entry.migrate = config.entry.main[0].replace('main.ts', 'migrate')
+  const basePath = 'apps/services/search-indexer/src'
+  config.entry = {
+    ...config.entry,
+    migrateAws: './' + path.join(basePath, './migrate/migrateAws.ts'),
+    migrateElastic: './' + path.join(basePath, './migrate/migrateElastic.ts'),
+    migrateKibana: './' + path.join(basePath, './migrate/migrateKibana.ts'),
+  }
   config.output.filename = '[name].js'
   return {
     ...config,
