+++ workspace.json
@@ -4809,8 +4809,12 @@
         "migrate": {
           "builder": "@nrwl/workspace:run-commands",
           "options": {
-            "command": "yarn ts-node -P apps/services/search-indexer/tsconfig.app.json apps/services/search-indexer/src/migrate/index.ts",
-            "cwd": "apps/services/search-indexer"
+            "commands": [
+              "yarn ts-node -P apps/services/search-indexer/tsconfig.app.json apps/services/search-indexer/src/migrate/migrateElastic.ts",
+              "yarn ts-node -P apps/services/search-indexer/tsconfig.app.json apps/services/search-indexer/src/migrate/migrateKibana.ts"
+            ],
+            "cwd": "apps/services/search-indexer",
+            "parallel": false
           }
         },
         "lint": {
